requirements.txt
torch
torchvision
fastapi
uvicorn
opencv-python
numpy
pillow
tqdm
python-multipart
python-dotenv
uvicorn

